pyproject.toml
[tool.poetry]
name = "game-of-life-pvp"
version = "0.1.0"
description = "Conway's Game of Life PvP, inspired by DougDoug"
authors = ["Florian Bender <benderflorian70@gmail.com>"]
readme = "README.md"
packages = [{ include = "game_of_life_pvp" }]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
numpy = "^1.26.0"
pillow = "^10.0.1"
mypy = "^1.5.1"


[tool.poetry.group.dev.dependencies]
black = "^23.9.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
